requirements.txt
Django==4.1.4
django-phonenumber-field==7.0.1
phonenumbers==8.13.2
python-dotenv==0.20.*
python-telegram-bot==13.*
